+++ src/doti/helpers/boolean.py
@@ -27,3 +27,8 @@ def is_true(string):
     """Check if a string represents a true value."""
     bools = ["add", "stow", "true", "yes", "on", "1"]
     return string.lower() in bools
+
+def is_false(string):
+    """Check if a string represents a true value."""
+    bools = ["remove", "unstow", "false", "no", "off", "0"]
+    return string.lower() in bools