+++ src/doti/helpers/system.py
@@ -5,7 +5,7 @@ from os import environ
 from platform import node, system
 
 
-def get_system():
+def get_platform():
     """Returns platform name."""
     if "TERMUX_VERSION" in environ:
         platform = "termux"
@@ -26,6 +26,18 @@ def get_system():
 
     return platform
 
+def get_distro():
+    """Returns distro name."""
+    try:
+        with open("/etc/os-release") as file:
+            for line in file:
+                key_value = line.partition("=")
+                if key_value[0] == "ID":
+                    distro = key_value[-1].rstrip()
+    except FileNotFoundError:
+        distro = "unknown_distro"
+    
+    return distro
 
 def get_hostname():
     """Returns hostname."""